feat(tasks logs) add logging interceptor refactor(auth db logs)

+++ tasks/internal/service/tasks.go
@@ -4,9 +4,11 @@ import (
 	"context"
 	"errors"
 	"log/slog"
+	"strings"
 
 	pb "github.com/Novip1906/tasks-grpc/tasks/api/proto/gen"
 	"github.com/Novip1906/tasks-grpc/tasks/internal/config"
+	"github.com/Novip1906/tasks-grpc/tasks/internal/contextkeys"
 	"github.com/Novip1906/tasks-grpc/tasks/internal/models"
 	"github.com/Novip1906/tasks-grpc/tasks/internal/storage"
 	"github.com/Novip1906/tasks-grpc/tasks/pkg/logging"
@@ -34,30 +36,27 @@ func NewTasksService(config *config.Config, log *slog.Logger, db TasksStorage) *
 }
 
 func (s *TasksService) CreateTask(ctx context.Context, req *pb.CreateTaskRequest) (*pb.CreateTaskResponse, error) {
-	userId, ok := ctx.Value(models.UserIDContextKey).(int64)
+	userId, ok := contextkeys.GetUserId(ctx)
 	if !ok {
-		return nil, status.Error(codes.Internal, "!ok")
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
+	log := contextkeys.GetLogger(ctx)
 
 	text := req.GetText()
 
-	log := s.log.With(
-		slog.String("op", "Tasks.CreateTask"),
-		slog.Int64("user_id", userId),
-		slog.String("text", text),
-	)
+	log.Debug("attempt")
 
-	log.Info("attempt")
+	text = processText(text)
 
 	if !textIsValid(text, s.cfg) {
 		log.Error("text len invalid")
-		return nil, status.Error(codes.InvalidArgument, "invalid text length")
+		return nil, status.Error(codes.InvalidArgument, ErrInvalidTextMessage)
 	}
 
 	err := s.db.CreateTask(userId, text)
 	if err != nil {
-		log.Error("db.CreateTask", logging.Err(err))
-		return nil, status.Error(codes.Internal, ErrInternal.Error())
+		log.Error("db error", logging.DbErr("CreateTask", err))
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
 
 	log.Info("task created")
@@ -67,51 +66,47 @@ func (s *TasksService) CreateTask(ctx context.Context, req *pb.CreateTaskRequest
 }
 
 func (s *TasksService) GetTask(ctx context.Context, req *pb.GetTaskRequest) (*pb.Task, error) {
-	userId := ctx.Value(models.UserIDContextKey).(int64)
+	userId, ok := contextkeys.GetUserId(ctx)
+	if !ok {
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
+	}
 	taskId := req.GetTaskId()
 
-	log := s.log.With(
-		slog.String("op", "Tasks.GetTask"),
-		slog.Int64("user_id", userId),
-		slog.Int64("task_id", taskId),
-	)
+	log := contextkeys.GetLogger(ctx).With(slog.Int64("task_id", taskId))
 
-	log.Info("attempt")
+	log.Debug("attempt")
 
 	task, err := s.db.GetTask(userId, taskId)
 	if errors.Is(err, storage.ErrTaskNotFound) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.NotFound, err.Error())
+		log.Error("task not found", logging.Err(err))
+		return nil, status.Error(codes.NotFound, ErrTaskNotFoundMessage)
 	}
 	if errors.Is(err, storage.ErrNotTaskAuthor) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.PermissionDenied, err.Error())
+		log.Error("user is not task author", logging.Err(err))
+		return nil, status.Error(codes.PermissionDenied, ErrNotTaskAuthorMessage)
 	}
 	if err != nil {
-		log.Error("db.GetTask", logging.Err(err))
-		return nil, status.Error(codes.Internal, ErrInternal.Error())
+		log.Error("db error", logging.DbErr("GetTask", err))
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
-
-	log.Info("completed")
 
 	return &pb.Task{Text: task.Text, AuthorName: task.AuthorName, CreatedAt: task.CreatedAt.Unix()}, nil
 
 }
 
 func (s *TasksService) GetAllTasks(ctx context.Context, req *pb.GetAllTasksRequest) (*pb.GetAllTasksResponse, error) {
-	userId := ctx.Value(models.UserIDContextKey).(int64)
+	userId, ok := contextkeys.GetUserId(ctx)
+	if !ok {
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
+	}
+	log := contextkeys.GetLogger(ctx)
 
-	log := s.log.With(
-		slog.String("op", "Tasks.GetAllTasks"),
-		slog.Int64("user_id", userId),
-	)
-
-	log.Info("get all tasks attempt")
+	log.Debug("get all tasks attempt")
 
 	tasks, err := s.db.GetAllTasks(userId)
 	if err != nil {
-		log.Error("db.GetAllTasks", logging.Err(err))
-		return nil, status.Error(codes.Internal, ErrInternal.Error())
+		log.Error("db error", logging.DbErr("GetAllTasks", err))
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
 
 	pbTasks := make([]*pb.Task, 0, len(tasks))
@@ -124,43 +119,41 @@ func (s *TasksService) GetAllTasks(ctx context.Context, req *pb.GetAllTasksReque
 		pbTasks = append(pbTasks, pbTask)
 	}
 
-	log.Info("completed")
-
 	return &pb.GetAllTasksResponse{Tasks: pbTasks}, nil
 
 }
 
 func (s *TasksService) UpdateTask(ctx context.Context, req *pb.UpdateTaskRequest) (*pb.UpdateTaskResponse, error) {
-	userId := ctx.Value(models.UserIDContextKey).(int64)
+	userId, ok := contextkeys.GetUserId(ctx)
+	if !ok {
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
+	}
 	taskId := req.GetTaskId()
 	newText := req.GetNewText()
 
-	log := s.log.With(
-		slog.String("op", "Tasks.UpdateTask"),
-		slog.Int64("user_id", userId),
-		slog.Int64("task_id", taskId),
-		slog.String("new_text", newText),
-	)
+	log := contextkeys.GetLogger(ctx).With(slog.Int64("task_id", taskId))
 
-	log.Info("attempt")
+	log.Debug("attempt")
+
+	newText = processText(newText)
 
 	if !textIsValid(newText, s.cfg) {
 		log.Error("text len invalid")
-		return nil, status.Error(codes.InvalidArgument, "invalid text length")
+		return nil, status.Error(codes.InvalidArgument, ErrInvalidTextMessage)
 	}
 
 	err := s.db.UpdateTask(userId, taskId, newText)
 	if errors.Is(err, storage.ErrTaskNotFound) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.NotFound, err.Error())
+		log.Error("task not found", logging.Err(err))
+		return nil, status.Error(codes.NotFound, ErrTaskNotFoundMessage)
 	}
 	if errors.Is(err, storage.ErrNotTaskAuthor) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.PermissionDenied, err.Error())
+		log.Error("user is not task author", logging.Err(err))
+		return nil, status.Error(codes.PermissionDenied, ErrNotTaskAuthorMessage)
 	}
 	if err != nil {
-		log.Error("db.UpdateTask", logging.Err(err))
-		return nil, status.Error(codes.Internal, ErrInternal.Error())
+		log.Error("db error", logging.DbErr("UpdateTask", err))
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
 
 	log.Info("task updated")
@@ -169,29 +162,28 @@ func (s *TasksService) UpdateTask(ctx context.Context, req *pb.UpdateTaskRequest
 }
 
 func (s *TasksService) DeleteTask(ctx context.Context, req *pb.DeleteTaskRequest) (*pb.DeleteTaskResponse, error) {
-	userId := ctx.Value(models.UserIDContextKey).(int64)
+	userId, ok := contextkeys.GetUserId(ctx)
+	if !ok {
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
+	}
 	taskId := req.GetTaskId()
 
-	log := s.log.With(
-		slog.String("op", "Tasks.CreateTask"),
-		slog.Int64("user_id", userId),
-		slog.Int64("task_id", taskId),
-	)
+	log := contextkeys.GetLogger(ctx).With(slog.Int64("task_id", taskId))
 
-	log.Info("attempt")
+	log.Debug("attempt")
 
 	err := s.db.DeleteTask(userId, taskId)
 	if errors.Is(err, storage.ErrTaskNotFound) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.NotFound, err.Error())
+		log.Error("task not found", logging.Err(err))
+		return nil, status.Error(codes.NotFound, ErrTaskNotFoundMessage)
 	}
 	if errors.Is(err, storage.ErrNotTaskAuthor) {
-		log.Error(err.Error())
-		return nil, status.Error(codes.PermissionDenied, err.Error())
+		log.Error("user is not task author", logging.Err(err))
+		return nil, status.Error(codes.PermissionDenied, ErrNotTaskAuthorMessage)
 	}
 	if err != nil {
-		log.Error("db.DeleteTask", logging.Err(err))
-		return nil, status.Error(codes.Internal, ErrInternal.Error())
+		log.Error("db error", logging.DbErr("DeleteTask", err))
+		return nil, status.Error(codes.Internal, ErrInternalMessage)
 	}
 
 	log.Info("task deleted")
@@ -201,4 +193,8 @@ func (s *TasksService) DeleteTask(ctx context.Context, req *pb.DeleteTaskRequest
 
 func textIsValid(text string, cfg *config.Config) bool {
 	return len(text) >= cfg.Params.Text.Min && len(text) <= cfg.Params.Text.Max
+}
+
+func processText(text string) string {
+	return strings.TrimSpace(text)
 }
